deps: Update superstring to commit c2ff06231736236
This revision makes superstring easily buildable on Windows as a git
dependency or tarball URL dependency, by vendoring in the 'win-iconv'
files as regular files, not as a submodule.

npm and Yarn do not necessarily install submodules of git repos as of
their latest versions. And tarballs wouldn't be generated with the
submodule content, as far as I know.

So, doing it this way is much more broadly compatible with how the
package managers tend to work these days.